delete unnecessary event fragment
